requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=0.32.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.4.0=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.2.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
datasets=2.20.0=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
dppy=0.3.2=pypi_0
evaluate=0.4.2=pypi_0
expat=2.6.2=h6a678d5_0
faiss-gpu=1.8.0=py3.12_hedc54c9_0_cuda11.4.4
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fonttools=4.53.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.5=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.7=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py312h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.10.1.7=0
libcurand=10.3.6.82=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libfaiss=1.8.0=h5aaf3ed_0_cuda11.4.4
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lxml=5.2.2=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.9.1=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py312h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
omegaconf=2.3.0=pypi_0
openai=1.37.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
packaging=24.1=py312h06a4308_0
pandas=2.2.2=pypi_0
peft=0.12.0=pypi_0
pillow=10.4.0=py312h5eee18b_0
pip=24.0=py312h06a4308_0
portalocker=2.10.1=pypi_0
psutil=6.0.0=pypi_0
pyarrow=17.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.4=h5148396_1
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.2.0=py3.12_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py312h5eee18b_0
rank-bm25=0.2.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.32.2=py312h06a4308_0
rouge-score=0.1.2=pypi_0
sacrebleu=2.4.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
setuptools=69.5.1=py312h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py312h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
torchaudio=2.2.0=py312_cu118
torchvision=0.17.0=py312_cu118
tqdm=4.66.4=pypi_0
transformers=4.43.3=pypi_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=py312h06a4308_0
wheel=0.43.0=py312h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
